reworking ability chat card
attack card has been rebranded to ability chat card and will show everything that has to do with an ability. It will have buttons that appear only if they are necessary. For example if an ability has an attack component then there will be an attack roll and damage button. Also buttons will only be visible to whoever rolled the ability and the GM